// Web/src/app/views/catalog/Models/Enums/RouteIds.ts
export enum RouteIds {
    Settings = 1018008,
    RawItem = 1019013,
    Salereport = 1019005,
    ChartOfAccount = 1019025,
    VoucherTypes = 1019026,
    GeneralLedger = 1019030,
    Vouchers = 1019028,
    FiscalYear = 1019039,
    NotificationTemplate = 1019029,
    Clients = 1019032,
    Roles = 1019033,
    Permissions = 1019036,
    Group = 1020001,
    Category = 1020002,
    SubCategory = 1020003,
    Supplier = 1020006,
    Brand = 1020008,
    Store = 1020009,
    Branch = 1020011,
    Size = 1020013,
    UOM = 1020034,
    UOMConverison = 1020035,
    MenuGroup = 1020037,
    MenuItem = 1020038,
    Currency = 1020039,
    PrinterSetting = 1020041,
    RecipeModifier = 1020042,
    GST = 1020043,
    DeviceSettings = 1020044,
    OnlineDBSetting = 1020044,
    CustomerInfo = 1020048,
    MemberPoints = 1020049,
    Modifier = 1020050,
    DiscountKeys = 1020051,
    ServiceCharges = 1020053,
    Banks = 1020054,
    LoyaltyCard = 1020055,
    DiscountCompaign = 1020057,
    OrderSource = 1020058,
    SubItems = 1020059,
    DiscVoucher = 1020060,
    Table = 1020061,
    AttachMenu = 1020062,
    Floor = 1020064,
    CompanyInfo = 1020022,
    Expanses = 1020023,
    Staff = 1020028,
    EmployeeSale = 1020029,
    Employee = 1020030,
    EmployeeSalary = 1020031,
    Attendance = 1020032,
    EmployeeStatement = 1020033,
    Users = 1020065,

    LogEvents = 1019037


}